# Small optimization in OptimizedScalarQuantizer by using mul instead of div